move AccessSniff to app file
More modular
